Python code to calculate area, circumference, and change radius of a circle using class.
This is a Python code that defines a Circle class that has methods to calculate the area and circumference of a circle, as well as to change the radius of the circle. The class takes the radius as an argument in its constructor method (__init__) and stores it as an instance variable.

The area method calculates the area of the circle using the formula πr², where r is the radius of the circle. It uses the pi constant from the math module and the ** operator to calculate the square of the radius.

The circum method calculates the circumference of the circle using the formula 2πr, where r is the radius of the circle. It also uses the pi constant from the math module.

The change_radius method allows the user to change the radius of the circle. It takes a new radius as an argument and sets the instance variable radius to the new value.

The code then creates an instance of the Circle class with a radius of 1 and enters an infinite loop that prompts the user to enter a new radius. It calls the change_radius method to update the radius of the circle and then prints out the updated area and circumference using the area and circum methods.